chat/requirements.txt
scikit-learn
tensorflow-gpu==2.3.1
tensorlayer
numpy
click
tqdm
nltk
